refactor: SavingsCalculator 폴더 분리 및 컴포넌트 분리
