+++ src/pages/savingsCalculator/index.tsx
@@ -1,10 +1,8 @@
-import { useQuery } from '@tanstack/react-query';
-import { savingsProductQueryOptions } from 'queries/savings/queries';
 import { ChangeEvent, Suspense, useState } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
-import { filterSavingsProducts } from 'service/savingsCalculator.service';
-import { Assets, Border, colors, ListRow, NavigationBar, SelectBottomSheet, Spacing, Tab, TextField } from 'tosslib';
-import { SavingsProduct } from 'types/savingsProducts';
+import { Border, NavigationBar, SelectBottomSheet, Spacing, Tab, TextField } from 'tosslib';
+
+import { SavingsProducts } from './components/SavingsProducts';
 
 export function SavingsCalculatorPage() {
   return (
@@ -15,49 +13,6 @@ export function SavingsCalculatorPage() {
     </ErrorBoundary>
   );
 }
-
-export interface Filter {
-  goalPrice: number;
-  monthlyAmount: number;
-  term: number;
-}
-const SavingsProducts = ({ filter }: { filter: Filter }) => {
-  const { data: savingsProducts } = useQuery(savingsProductQueryOptions());
-
-  if (!savingsProducts || savingsProducts.length === 0) {
-    return <div>적금 상품이 없습니다.</div>;
-  }
-
-  const filteredSavingsProducts = filterSavingsProducts(filter, savingsProducts);
-  if (!filteredSavingsProducts || filteredSavingsProducts.length === 0) {
-    return <div>조건에 맞는 적금 상품이 없습니다.</div>;
-  }
-
-  return (
-    <>
-      {filteredSavingsProducts.map((product: SavingsProduct) => {
-        return (
-          <ListRow
-            key={product.id}
-            contents={
-              <ListRow.Texts
-                type="3RowTypeA"
-                top={product.name}
-                topProps={{ fontSize: 16, fontWeight: 'bold', color: colors.grey900 }}
-                middle={`연 이자율: ${product.annualRate}%`}
-                middleProps={{ fontSize: 14, color: colors.blue600, fontWeight: 'medium' }}
-                bottom={`${product.minMonthlyAmount.toLocaleString()}원 ~ ${product.maxMonthlyAmount.toLocaleString()}원 | ${product.availableTerms}개월`}
-                bottomProps={{ fontSize: 13, color: colors.grey600 }}
-              />
-            }
-            right={<Assets.Icon name="icon-check-circle-green" />}
-            onClick={() => {}}
-          />
-        );
-      })}
-    </>
-  );
-};
 
 const SavingCalculator = () => {
   const [goalPrice, setGoalPrice] = useState<number>(0);
